colocate test files and impl, remove test globals file as not needed
