+++ server/meta.test.js
@@ -4,7 +4,7 @@
 import cheerio from "cheerio";
 import { test, afterEach } from "vitest";
 
-import { unpackSDKMeta } from "../../server";
+import { unpackSDKMeta } from ".";
 
 afterEach(() => {
   // eslint-disable-next-line no-process-env
