+++ src/tracking.test.js
@@ -1,7 +1,7 @@
 /* @flow */
 import { describe, it, vi, expect } from "vitest";
 
-import { getSDKInitTime, setupLogger } from "../../src/tracking";
+import { getSDKInitTime, setupLogger } from "./tracking";
 
 const clientId = "foobar123";
 const mockScriptSrc = `https://test.paypal.com/sdk/js?client-id=${clientId}`;
